Various 4.2.0 changes
Allow rendering the boss bar as a nameplate
Changed how particles spawn for new wave entities
Removed the default spacing rule for gateways
Added a count field to WaveEntity
Made it so that entities walking out of leash range are respawned instead of causing a failure
Removed `completion_xp` in favor of the `experience` reward type.
Made it so that fog can be disabled.
Made it so that Wave Entities can specify a custom description key.